Implemented user cart workers
We have added the ability to create new user carts and get them.

When we call the `ExCashier.start_user_cart/1` function, we are creating a new user cart. This, as it's a supervised process, will have an internal state where we will store the user's id and all the items added to its cart.

We also have two more functions:

- `ExCashier.add_item/2` - which allows us to add an item to the user's cart
- `ExCashier.get_cart/1` - that will return a user's cart

Notes:

- Added restriction for the user and item identifiers to be **strings**
- All the user cart processes are registered under the user identifier in the `ExCashier.UserCartRegistry`
- The previous restricts the number of carts per user to 1
- Added tests for the mentioned features
- Added some simple documentation and specs