Implemented the catalogue
The Catalogue will load and store the items specified in its origin file.

With that, we could implement modification functions for the saved products.

Also, modifier the `add_item/2` function to be `add_item/3`, adding the capacity of adding different quantity of the same item at a time.

Added more tests for the previous change.

Finally, added tests for the new feature. Getting a coverage of almost 100%.